Move the `load_workqueue` function to `pr_tracking`
