+++ src/handlers/pr_tracking.rs
@@ -130,6 +130,22 @@ pub(super) async fn handle_input<'a>(
     Ok(())
 }
 
+/// Loads the workqueue (mapping of open PRs assigned to users) from GitHub
+pub async fn load_workqueue(client: &Octocrab) -> anyhow::Result<ReviewerWorkqueue> {
+    tracing::debug!("Loading workqueue for rust-lang/rust");
+    let prs = retrieve_pull_request_assignments("rust-lang", "rust", &client).await?;
+
+    // Aggregate PRs by user
+    let aggregated: HashMap<UserId, HashSet<PullRequestNumber>> =
+        prs.into_iter().fold(HashMap::new(), |mut acc, (user, pr)| {
+            let prs = acc.entry(user.id).or_default();
+            prs.insert(pr);
+            acc
+        });
+    tracing::debug!("PR assignments\n{aggregated:?}");
+    Ok(ReviewerWorkqueue::new(aggregated))
+}
+
 /// Retrieve tuples of (user, PR number) where
 /// the given user is assigned as a reviewer for that PR.
 /// Only non-draft, non-rollup and open PRs are taken into account.
