+++ src/handlers/pull_requests_assignment_update.rs
@@ -1,10 +1,6 @@
-use crate::github::PullRequestNumber;
-use crate::github::UserId;
-use crate::handlers::pr_tracking::{retrieve_pull_request_assignments, ReviewerWorkqueue};
+use crate::handlers::pr_tracking::load_workqueue;
 use crate::jobs::Job;
 use async_trait::async_trait;
-use octocrab::Octocrab;
-use std::collections::{HashMap, HashSet};
 
 pub struct PullRequestAssignmentUpdate;
 
@@ -22,20 +18,4 @@ impl Job for PullRequestAssignmentUpdate {
 
         Ok(())
     }
-}
-
-/// Loads the workqueue (mapping of open PRs assigned to users) from GitHub
-pub async fn load_workqueue(client: &Octocrab) -> anyhow::Result<ReviewerWorkqueue> {
-    tracing::debug!("Loading workqueue for rust-lang/rust");
-    let prs = retrieve_pull_request_assignments("rust-lang", "rust", &client).await?;
-
-    // Aggregate PRs by user
-    let aggregated: HashMap<UserId, HashSet<PullRequestNumber>> =
-        prs.into_iter().fold(HashMap::new(), |mut acc, (user, pr)| {
-            let prs = acc.entry(user.id).or_default();
-            prs.insert(pr);
-            acc
-        });
-    tracing::debug!("PR assignments\n{aggregated:?}");
-    Ok(ReviewerWorkqueue::new(aggregated))
 }
